+++ src/pages/Dashboard/DashboardSidebar.tsx
@@ -1,7 +1,7 @@
 import { Stack } from '@chakra-ui/react'
 
 import { RecentTransactions } from './RecentTransactions'
-import { TradeCard } from './TradCard'
+import { TradeCard } from './TradeCard'
 
 export const DashboardSidebar = () => {
   return (
