+++ src/pages/Dashboard/TradeCard.tsx
